requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
jinja2>=3.1.3
pymupdf>=1.23.0
Pillow>=10.0.0
PyYAML>=6.0
